pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = "migrations/"

[tool.isort]
skip_glob = ["__pycache__", "**/migrations"]
profile = "black"

[tool.poetry]
name = "neatplus-server"
version = "0.0.1"
description = "Neatplus server"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
environs = {version = "^9.3.2", extras = ["django"]}
marshmallow = "^3.11.1"
uvicorn = {version = "^0.13.4", extras = ["standard"], optional=true}
gunicorn = {version="^20.1.0", optional=true}
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"
django-filter = "^2.4.0"
djangorestframework-simplejwt = "^4.6.0"
djangorestframework-camel-case = "^1.2.0"
django-admin-interface = "^0.16.1"
django-otp = "^1.0.3"
qrcode = "^6.1"
django-cors-headers = "^3.7.0"
django-silk = "^4.1.0"
django-storages = "^1.11.1"
sentry-sdk = "^1.0.0"
celery = "^5.0.5"
redis = "^3.5.3"
django-modeltranslation = "^0.16.2"
django-simple-history = "^2.12.0"
django-ordered-model = "^3.4.1"
pytz = "^2021.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
safety = "^1.10.3"
pre-commit = "^2.12.0"
django-extensions = "^3.1.3"
ptpython = "^3.0.17"
ipython = "^7.22.0"
Werkzeug = "^1.0.1"
pyparsing = "^2.4.7"
pydot = "^1.4.2"

[tool.poetry.extras]
asgi = ["uvicorn", "gunicorn"]
